add get_code_blobs() to mcwamp.cpp
This functionality used to be part of read_code_bundles(), but was pulled out
so it could be called separately.

It is called when HCC_LAZYINIT is OFF so that the same code blobs that would
have been loaded by a non-lazy init are still loaded, but not yet made into
code bundles.